pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-bot"
version = "0.1.0"
description = "CryptoBot for different profiles : investors, journalist and traders"
requires-python = ">=3.12"
dependencies = [
  "ccxt>=4.0.0",
  "click>=8.1.7",
  "python-dotenv>=1.0.0",
  "pandas>=2.0.0",
  "numpy>=1.26.0",
  "schedule>=1.2.0",
  "psycopg2-binary>=2.9.9",
  "SQLAlchemy>=2.0.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "polars>=1.37.1",
]

[dependency-groups]
dev = [
  "pytest ~=8.3",
  "pytest-cov ~=5.0",
  "ruff ~=0.5",
  "mypy ~=1.10",
  "pre-commit ~=3.8",
]


[project.scripts]
crypto-bot = "src.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.editable]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".venv", "venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_equality = true
exclude = ["src/tests/", "venv", ".venv", "build", "dist", "src/register/_refactor"]
ignore_missing_imports = true



[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q"
